feat/code reorganization into multiple smaller files
