+++ index.js
@@ -1,42 +1,20 @@
 /* eslint-disable max-len */
 const axiosRaw = require('axios');
-const Promise = require('bluebird');
 const R = require('ramda');
 const assert = require('assert');
 const moment = require('moment');
 const js2xmlparser = require('js2xmlparser');
 const xml2js = require('xml2js');
 const { XMLParser } = require('fast-xml-parser');
-const { translateTPOption } = require('./resolvers/product');
-const { translateItineraryBooking } = require('./resolvers/itinerary');
-
 const Normalizer = require('./normalizer');
 
-const DEFAULT_TOURPLAN_SERVICE_STATUS = 'IR';
-const DEFAULT_CUSTOM_RATE_MARKUP_PERCENTAGE = 0;
-const MIN_MARKUP_PERCENTAGE = 1;
-const MAX_MARKUP_PERCENTAGE = 100;
-const MIN_EXTENDED_BOOKING_YEARS = 1;
-const MAX_EXTENDED_BOOKING_YEARS = 100;
-const GENERIC_AVALABILITY_CHK_ERROR_MESSAGE = 'Not bookable for the requested date/stay. (e.g. no rates, block out period, on request, minimum stay etc.)';
-const SERVICE_CANNOT_BE_ADDED_ERROR_MESSAGE = 'Service cannot be added to quote for the requested date/stay. (e.g. no rates, block out period, on request, minimum stay etc.)';
-const MAX_PAX_EXCEEDED_ERROR_TEMPLATE = 'Maximum {maxPax} pax allowed per Pax Config. Please update the Pax Config accordingly.';
-const RATES_AVAILABLE_TILL_ERROR_TEMPLATE = 'Rates are only available until {dateTill}. Please change the date and try again.';
-const RATES_CLOSED_ERROR_TEMPLATE = 'The rates are closed for the given dates: {closedDateRanges}. Please try again with a different dates range.';
-const MIN_STAY_LENGTH_ERROR_TEMPLATE = '{minSCUDateRangesText}. Please adjust the stay length and try again.';
-const USER_FRIENDLY_DATE_FORMAT = 'DD-MMM-YYYY';
-const DEFAULT_CUSTOM_RATES_EXTENDED_BOOKING_YEARS = 2;
-const EXTENDED_BOOKING_YEARS_ERROR_TEMPLATE = 'Last available rate until: {lastRateEndDate}. Custom rates can only be extended by {extendedBookingYears} year(s), please change the date and try again.';
-const CUSTOM_RATE_ID_NAME = 'Custom';
+const { searchAvailabilityForItinerary } = require('./api/availability/availability');
+const { addServiceToItinerary } = require('./api/add-service');
+const { searchProductsForItinerary } = require('./api/search-products-itinerary');
+const { searchItineraries } = require('./api/search-itinerary');
 
 const xmlParser = new xml2js.Parser();
 const fastParser = new XMLParser();
-const BAD_XML_CHARS = /[\u0000-\u0008\u000B\u000C\u000E-\u001F\u007f-\u0084\u0086-\u009f\uD800-\uDFFF\uFDD0-\uFDFF\uFFFF\uC008\uFEFF\u00DF]/g;
-const passengerTypeMap = {
-  Adult: 'Adult',
-  Child: 'Child',
-  Infant: 'Infant',
-};
 
 const defaultXmlOptions = {
   prettyPrinting: { enabled: false },
@@ -45,23 +23,12 @@ const defaultXmlOptions = {
     name: 'tourConnect_4_00_000.dtd',
   },
 };
-const hostConnectXmlOptions = {
-  prettyPrinting: { enabled: false },
-  dtd: {
-    include: true,
-    name: 'hostConnect_4_06_009.dtd',
-  },
-};
 const getHeaders = ({ length }) => ({
   Accept: 'application/xml',
   'Content-Type': 'application/xml; charset=utf-8',
   'Content-Length': length,
 });
-const wildcardMatch = (wildcard, str) => {
-  const w = wildcard.replace(/\s/g, '').replace(/[.+^${}()|[\]\\]/g, '\\$&'); // regexp escape
-  const re = new RegExp(`${w.replace(/\*/g, '.*').replace(/\?/g, '.')}`, 'i');
-  return re.test(str.replace(/\s/g, ''));
-};
+
 class BuyerPlugin {
   constructor(params = {}) { // we get the env variables from here
     Object.entries(params).forEach(([attr, value]) => {
@@ -360,863 +327,6 @@ class BuyerPlugin {
       }
       return R.path(['Reply'], replyObj);
     };
-
-    this.getRoomConfigs = (paxConfigs, noPaxList) => {
-      // There should be only 1 RoomConfigs for AddServiceRequest
-      let RoomConfigs = {};
-      // add one RoomConfig for each room required (i.e. one for each PaxConfig)
-      RoomConfigs.RoomConfig = [];
-      let indexRoomConfig = 0;
-      paxConfigs.forEach(({ roomType, adults, children, infants, passengers = [] }) => {
-        const EachRoomConfig = passengers.length ? passengers.reduce((acc, p) => {
-          if (p.passengerType === passengerTypeMap.Adult) {
-            acc.Adults += 1;
-          }
-          if (p.passengerType === passengerTypeMap.Child) {
-            acc.Children += 1;
-          }
-          if (p.passengerType === passengerTypeMap.Infant) {
-            acc.Infants += 1;
-          }
-          return acc;
-        }, {
-          Adults: 0,
-          Children: 0,
-          Infants: 0,
-        }) : {
-          Adults: adults || 0,
-          Children: children || 0,
-          Infants: infants || 0,
-        };
-        const RoomType = ({
-          Single: 'SG',
-          Double: 'DB',
-          Twin: 'TW',
-          Triple: 'TR',
-          Quad: 'QD',
-          Other: 'OT',
-        })[roomType];
-        if (RoomType) EachRoomConfig.RoomType = RoomType;
-        if (passengers && passengers.length && !noPaxList) {
-          // There should be only 1 PaxList inside each EachRoomConfig
-          EachRoomConfig.PaxList = {};
-          // Inside PaxList, there should be 1 PaxDetail for each passenger (Pax)
-          EachRoomConfig.PaxList.PaxDetails = passengers.map(p => {
-            /*
-              TP API doesn't allow us to modify existing pax details
-              when PersonId is present, other details are ignored by TP anyways
-              when it is not present, TP is comparing every key in PaxDetail to identify
-              duplicate, so if we send Pax Detail with the same first and last name, but different
-              Age, TP will consider them to be different pax, which actually is duplicate, given
-              sometimes AI could be extracting inconsistent data
-            */
-            if (p.personId) {
-              return {
-                PersonId: p.personId,
-              };
-            }
-            const EachPaxDetails = {
-              Forename: this.escapeInvalidXmlChars(p.firstName),
-              Surname: this.escapeInvalidXmlChars(p.lastName),
-              PaxType: {
-                Adult: 'A',
-                Child: 'C',
-                Infant: 'I',
-              }[p.passengerType] || 'A',
-            };
-            if (p.salutation) EachPaxDetails.Title = this.escapeInvalidXmlChars(p.salutation);
-            if (p.dob) EachPaxDetails.DateOfBirth = p.dob;
-            // NOTE: TourPlan API doesn't accept age as empty string, i.e. empty XML tag <Age/>
-            // and trhows and error like - "1000 SCN System.InvalidOperationException: There is an
-            // error in XML document (29, 8). (Input string was not in a correct format.)"
-            // The solution is to NOT send the Age tag if it's empty
-            if (!R.isNil(p.age) && !Number.isNaN(p.age) && p.age) {
-              if (!(p.passengerType === passengerTypeMap.Adult && p.age === 0)) {
-                EachPaxDetails.Age = p.age;
-              }
-            }
-            return EachPaxDetails;
-          });
-        }
-        RoomConfigs.RoomConfig[indexRoomConfig++] = EachRoomConfig;
-      });
-      return RoomConfigs;
-    };
-    this.escapeInvalidXmlChars = str => {
-      if (!str) return '';
-      const convertAccentedChars = s => {
-        // according to TC-143, we will go through one mapping first
-        // for certain accented chars
-        const accentedChars = [
-          ['Ä', 'Ae'],
-          ['Ö', 'Oe'],
-          ['Ü', 'Ue'],
-          ['Ø', 'Oe'],
-          ['Å', 'Aa'],
-          ['Æ', 'Ae'],
-          ['ä', 'ae'],
-          ['ö', 'oe'],
-          ['ü', 'ue'],
-          ['ø', 'oe'],
-          ['å', 'aa'],
-          ['æ', 'ae'],
-          ['ß', 'ss'],
-        ];
-        const preprocessed = accentedChars.reduce((acc, [k, v]) => acc.replace(k, v), s);
-        return preprocessed.normalize('NFD').replace(/[\u0300-\u036f]/g, '');
-      };
-      // NOTE: There is no need to sanitize the string for 5 characters (&, <, >, " and ')
-      // because js2xmlparser does that for us. Plus if we use sanitize before calling js2xmlparser
-      // js2xmlparser will escape & to '&amp;' making it invalid XML
-      return convertAccentedChars(str)
-        .replace(/’/g, "'")
-        .replace(/‘/g, "'")
-        .replace(/“/g, '"')
-        .replace(/”/g, '"')
-        .replace(/–/g, '-')
-        .replace(BAD_XML_CHARS, '');
-    };
-
-    this.extractCancelPolicies = (rate, path, isOptionCancelPolicy) => {
-      const rawPolicies = R.pathOr([], path, rate);
-      let policies = [];
-      if (!Array.isArray(rawPolicies)) {
-        policies = [rawPolicies]; // If single item, convert to array
-      } else {
-        policies = rawPolicies;
-      }
-      return policies.map(policy => {
-        const mappedPolicy = {
-          // The description of the penalty (optional)
-          penaltyDescription: R.path(['PenaltyDescription'], policy),
-          // The number of OffsetTimeUnit in this relative deadline
-          cancelNum: R.path(['Deadline', 'OffsetUnitMultiplier'], policy),
-          // One of Second, Hour, Day, Week, Month or Year
-          cancelTimeUnit: R.path(['Deadline', 'OffsetTimeUnit'], policy),
-        };
-
-        if (isOptionCancelPolicy) {
-          // The absolute deadline, i.e. the final date and time of the deadline. (optional)
-          mappedPolicy.deadlineDateTime = R.path(['Deadline', 'DeadlineDateTime'], policy);
-          // Y if this penalty is the one used if a service line is cancelled now (N otherwise)
-          mappedPolicy.inEffect = R.path(['InEffect'], policy) && R.path(['InEffect'], policy) === 'Y';
-          // Amount of the cancellation penalty
-          mappedPolicy.cancelFee = R.path(['LinePrice'], policy);
-          // Line price less commission.
-          mappedPolicy.agentPrice = R.path(['AgentPrice'], policy);
-        }
-        return Object.fromEntries(
-          Object.entries(mappedPolicy).filter(([_, value]) => value !== undefined),
-        );
-      });
-    };
-
-    /*
-      Convert children or infants in the pax configs to adults.
-      Reason: Tourplan API doesn't support children or infants in the availability check
-      when ChildrenAllowed = N & CountChildrenInPaxBreak = Y or InfantsAllowed = N & CountInfantsInPaxBreak = Y.
-
-      Example:
-      const originalPaxConfigs = [
-        { roomType: 'DB', adults: 2, children: 1, infants: 1 },
-        { roomType: 'TW', adults: 1, children: 2, infants: 2 }
-      ];
-
-      OUTPUT:
-      const convertedPaxConfigs = convertToAdult(originalPaxConfigs, passengerTypeMap.Child);
-      Result: [
-        { roomType: 'DB', adults: 3, children: 0, infants: 1 },
-        { roomType: 'TW', adults: 2, children: 0, infants: 2 }
-      ]
-      OR
-      const convertedPaxConfigs = convertToAdult(originalPaxConfigs, passengerTypeMap.Infant);
-      Result: [
-        { roomType: 'DB', adults: 3, children: 1, infants: 0 },
-        { roomType: 'TW', adults: 2, children: 2, infants: 0 }
-      ]
-      NOTE: This method only converts children or infants to adults (based on the type parameter),
-      it doesn't convert the other type to adults.
-    */
-    this.convertToAdult = (paxConfigs, type) => {
-      if (!Array.isArray(paxConfigs)) {
-        return paxConfigs;
-      }
-
-      return paxConfigs.map(paxConfig => {
-        const newPaxConfig = { ...paxConfig };
-
-        // Convert children or infants to adults
-        const typeKey = type === passengerTypeMap.Child ? 'children' : 'infants';
-        const totalAdults = (newPaxConfig.adults || 0) + (newPaxConfig[typeKey] || 0);
-        newPaxConfig.adults = totalAdults;
-        newPaxConfig[typeKey] = 0;
-
-        // now update the type of the passengers to adults
-        if (newPaxConfig.passengers && Array.isArray(newPaxConfig.passengers)) {
-          newPaxConfig.passengers = newPaxConfig.passengers.map(passenger => ({
-            ...passenger,
-            passengerType: passenger.passengerType === type ? passengerTypeMap.Adult : passenger.passengerType,
-          }));
-        }
-
-        return newPaxConfig;
-      });
-    };
-
-    /*
-      Calculate the end date for an option based on start date, duration, and charge unit quantity.
-
-      @param {string} startDate - The start date in YYYY-MM-DD format
-      @param {number|null} duration - The duration in days (optional)
-      @param {number|null} chargeUnitQuantity - The number of charge units (optional)
-      @returns {string|null} The end date in YYYY-MM-DD format or null if not applicable
-    */
-    this.calculateEndDate = (startDate, duration, chargeUnitQuantity) => {
-      const startMoment = moment(startDate, 'YYYY-MM-DD');
-      let endDate = null;
-
-      if (duration) {
-        endDate = startMoment.clone().add(duration, 'days');
-      } else if (chargeUnitQuantity && chargeUnitQuantity > 1) {
-        endDate = startMoment.clone().add(chargeUnitQuantity, 'days');
-      }
-
-      return endDate ? endDate.format('YYYY-MM-DD') : null;
-    };
-
-    /*
-      Get the message for an option based on duration, charge unit quantity, and charge unit.
-
-      @param {number|null} duration - The duration in days (optional)
-      @param {number|null} chargeUnitQuantity - The number of charge units (optional)
-      @param {string|null} chargeUnit - The charge unit type (optional)
-      @returns {string|null} The message or null if not applicable
-    */
-    this.getOptionMessage = (duration, chargeUnitQuantity, chargeUnit) => {
-      const chargeUnitText = chargeUnit || 'Nights/Days';
-      let message = null;
-      if (duration && chargeUnitQuantity && duration !== chargeUnitQuantity) {
-        message = `This option allows exactly ${duration} ${chargeUnitText}. The end date is adjusted accordingly.`;
-      }
-      return message;
-    };
-
-    /**
-     * Validate that each RoomConfig does not exceed maxPaxPerCharge.
-     * Reason: Tourplan availability check returns success even if the pax configs exceed the maxPaxPerCharge.
-     * And then when the booking is made, the booking fails with an error like "002 SCN adults + children exceeds capacity".
-     *
-     * @param {Object} params - Validation parameters
-     * @returns {Object|null} Returns error object if validation fails, null if valid
-     */
-    this.validateMaxPaxPerCharge = ({
-      roomConfigs,
-      maxPaxPerCharge,
-    }) => {
-      // Verify that each RoomConfig does not exceed maxPaxPerCharge
-      if (maxPaxPerCharge && maxPaxPerCharge > 1) {
-        for (let i = 0; i < roomConfigs.RoomConfig.length; i++) {
-          const room = roomConfigs.RoomConfig[i];
-          const roomPax = (room.Adults || 0) + (room.Children || 0) + (room.Infants || 0);
-          if (roomPax > maxPaxPerCharge) {
-            // NOTE: As a long term solution, we need to return the errors per pax config
-            // so that the UI can display the errors for the particular pax config.
-            // For now we return on the 1st error and show the error in availability check.
-            return {
-              bookable: false,
-              type: 'inventory',
-              rates: [],
-              message: MAX_PAX_EXCEEDED_ERROR_TEMPLATE.replace('{maxPax}', maxPaxPerCharge),
-            };
-          }
-        }
-      }
-
-      return null;
-    };
-
-    /**
-     * Validates date ranges and room configurations
-     * @param {Object} params - Validation parameters
-     * @returns {Object|null} Returns error object if validation fails, null if valid
-     */
-    this.validateDateRanges = ({
-      dateRanges,
-      startDate,
-      chargeUnitQuantity,
-    }) => {
-      // Check if any rate set is closed
-      if (dateRanges.some(dateRange => dateRange.isClosed === 'Y')) {
-        const closedDateRanges = dateRanges.filter(dateRange => dateRange.isClosed === 'Y');
-        const closedDateRangesText = closedDateRanges.map(dateRange => {
-          const formattedStartDate = moment(dateRange.startDate).format(USER_FRIENDLY_DATE_FORMAT);
-          const formattedEndDate = moment(dateRange.endDate).format(USER_FRIENDLY_DATE_FORMAT);
-          return `${formattedStartDate} to ${formattedEndDate}`;
-        }).join(', ');
-        return {
-          bookable: false,
-          type: 'inventory',
-          rates: [],
-          message: RATES_CLOSED_ERROR_TEMPLATE.replace('{closedDateRanges}', closedDateRangesText),
-        };
-      }
-
-      // Check if any rate set has a minimum stay length
-      if (dateRanges.some(dateRange => dateRange.minSCU > 1)) {
-        const dateRangesWithMinSCUGreaterThanOne = dateRanges.filter(dateRange => dateRange.minSCU > 1);
-        const minSCUDateRangesText = [];
-
-        dateRangesWithMinSCUGreaterThanOne.forEach(dateRange => {
-          const daysBeforeDateRange = moment(dateRange.startDate).diff(moment(startDate), 'days');
-          const daysAfterDateRange = chargeUnitQuantity - daysBeforeDateRange;
-          if (daysAfterDateRange < dateRange.minSCU) {
-            const formattedStartDate = moment(dateRange.startDate).format(USER_FRIENDLY_DATE_FORMAT);
-            const formattedEndDate = moment(dateRange.endDate).format(USER_FRIENDLY_DATE_FORMAT);
-            minSCUDateRangesText.push(`The date range ${formattedStartDate} to ${formattedEndDate} has a minimum stay length of ${dateRange.minSCU}`);
-          }
-        });
-
-        if (minSCUDateRangesText.length > 0) {
-          return {
-            bookable: false,
-            type: 'inventory',
-            rates: [],
-            message: MIN_STAY_LENGTH_ERROR_TEMPLATE.replace('{minSCUDateRangesText}', minSCUDateRangesText.join(', ')),
-          };
-        }
-      }
-
-      return null; // No validation errors
-    };
-
-    /*
-      Get general option information from Tourplan API.
-
-      @param {string} optionId - The option ID to get information for
-      @param {string} hostConnectEndpoint - The HostConnect endpoint
-      @param {string} hostConnectAgentID - The agent ID
-      @param {string} hostConnectAgentPassword - The agent password
-      @param {Object} axios - The axios instance
-      @param {string} startDate - The start date
-      @param {number} chargeUnitQuantity - The number of charge units
-      @returns {Object} Object containing general option information
-    */
-    this.getGeneralAndDateRangesInfo = async (optionId, hostConnectEndpoint, hostConnectAgentID, hostConnectAgentPassword, axios, startDate, chargeUnitQuantity) => {
-      const getGeneralModel = checkType => ({
-        OptionInfoRequest: {
-          Opt: optionId,
-          Info: checkType,
-          AgentID: hostConnectAgentID,
-          Password: hostConnectAgentPassword,
-          DateFrom: startDate,
-          SCUqty: chargeUnitQuantity,
-        },
-      });
-
-      // Use G (General) & D (Date Ranges) check type to get the option general information
-      // and date ranges information
-      const replyObj = await this.callTourplan({
-        model: getGeneralModel('GD'),
-        endpoint: hostConnectEndpoint,
-        axios,
-        xmlOptions: hostConnectXmlOptions,
-      });
-      const GDCheck = R.path(['OptionInfoReply', 'Option'], replyObj);
-
-      const OptGeneralResult = R.pathOr({}, ['OptGeneral'], GDCheck);
-      const OptDateRangesResult = R.pathOr({}, ['OptDateRanges'], GDCheck);
-      const dateRanges = this.parseDateRanges(OptDateRangesResult);
-      const countChildrenInPaxBreak = R.pathOr(false, ['CountChildrenInPaxBreak'], OptGeneralResult) === 'Y';
-      const childrenAllowed = R.pathOr(false, ['ChildrenAllowed'], OptGeneralResult) === 'Y';
-      const infantsAllowed = R.pathOr(false, ['InfantsAllowed'], OptGeneralResult) === 'Y';
-      const countInfantsInPaxBreak = R.pathOr(false, ['CountInfantsInPaxBreak'], OptGeneralResult) === 'Y';
-      const duration = R.pathOr(null, ['Periods'], OptGeneralResult);
-      /* Charging multiple:
-        As per hostconnect documentation, this field is reported for non-accommodation and apartment
-        options (SType is N or A). However, test cases show that it is reported for accommodation
-        options as well. (SType is Y)
-        1.For apartments it gives the maximum number of adults that one apartment can hold.
-        2.For accommodation options, the same logic applies as for apartments. e.g. entire logde
-        3.For non-accommodation options, if MPFCU has a value of 1 then rates for the option are per-person.
-        MPFCU is greater than one then rates for this option are for a group, and MPFCU is the maximum
-        number of people (adults plus children) that can be booked per AddService call for the option.
-        Hence we need to check if the number of people in the roomConfigs is greater than maxPaxPerCharge.
-        A rental car might have an MPFCU of 4, for example.
-        NOTE: It is possible that we may have to revisit this for cases where the value is 1 (per-person)
-        and could apply to all types of services.
-      */
-      const maxPaxPerCharge = R.pathOr(null, ['MPFCU'], OptGeneralResult);
-      const chargeUnit = R.pathOr(null, ['SCU'], OptGeneralResult);
-
-      return {
-        countChildrenInPaxBreak,
-        childrenAllowed,
-        infantsAllowed,
-        countInfantsInPaxBreak,
-        duration,
-        maxPaxPerCharge,
-        chargeUnit,
-        dateRanges,
-      };
-    };
-
-    /*
-      Get date ranges for an option based on start date, charge unit quantity, and room configs.
-
-      @param {string} optionId - The option ID to get information for
-      @param {string} hostConnectEndpoint - The HostConnect endpoint
-      @param {string} hostConnectAgentID - The agent ID
-      @param {string} hostConnectAgentPassword - The agent password
-      @param {Object} axios - The axios instance
-      @param {string} startDate - The start date
-      @param {number} chargeUnitQuantity - The number of charge units
-      @param {Object} roomConfigs - The room configurations
-      @returns {Object} Object containing general option information
-    */
-    this.getOptionDateRanges = async (optionId, hostConnectEndpoint, hostConnectAgentID, hostConnectAgentPassword, axios, startDate, chargeUnitQuantity, roomConfigs) => {
-      const getDateRangesModel = checkType => ({
-        OptionInfoRequest: {
-          Opt: optionId,
-          Info: checkType,
-          AgentID: hostConnectAgentID,
-          Password: hostConnectAgentPassword,
-          DateFrom: startDate,
-          SCUqty: chargeUnitQuantity,
-          RoomConfig: roomConfigs,
-        },
-      });
-
-      // Use D (Date Ranges) check type to get the option date ranges information
-      const replyObj = await this.callTourplan({
-        model: getDateRangesModel('D'),
-        endpoint: hostConnectEndpoint,
-        axios,
-        xmlOptions: hostConnectXmlOptions,
-      });
-      const DCheck = R.path(['OptionInfoReply', 'Option'], replyObj);
-      const OptDateRangesResult = R.pathOr({}, ['OptDateRanges'], DCheck);
-      const dateRanges = this.parseDateRanges(OptDateRangesResult);
-
-      return dateRanges;
-    };
-
-    // Get the immediate last date range
-    // Logic: starting for the requested end date, keep going back in time until a rate range is found
-    this.getImmediateLastDateRange = async (optionId, hostConnectEndpoint, hostConnectAgentID, hostConnectAgentPassword, axios, endDate, roomConfigs) => {
-      // Prevent a very long period by limiting the number of days
-      const unitQuantity = Math.min(MAX_EXTENDED_BOOKING_YEARS * 365, moment(endDate).diff(moment(), 'days'));
-      const dateFrom = moment().format('YYYY-MM-DD');
-      const results = await this.getOptionDateRanges(optionId, hostConnectEndpoint, hostConnectAgentID, hostConnectAgentPassword, axios, dateFrom, unitQuantity, roomConfigs);
-      if (results.length > 0) {
-        return results[results.length - 1];
-      }
-
-      // If no date ranges found return null
-      return null;
-    };
-
-    /*
-      Create modified passenger configurations based on the count flags.
-      This function handles the conversion of children and infants to adults
-      when the respective count flags are enabled, which is necessary for
-      Tourplan API compatibility, see comments for treatChildrenAsAdults
-      and treatInfantsAsAdults for more details.
-    */
-    this.getModifiedPaxConfigs = (countChildrenInPaxBreak, childrenAllowed, countInfantsInPaxBreak, infantsAllowed, paxConfigs) => {
-      let modifiedPaxConfigs = paxConfigs;
-      if (countChildrenInPaxBreak && !childrenAllowed) {
-        // NOTE: If children are allowed let the availaiblity check happen with children
-        modifiedPaxConfigs = this.convertToAdult(paxConfigs, passengerTypeMap.Child);
-      }
-      if (countInfantsInPaxBreak && !infantsAllowed) {
-        // NOTE: If infants are allowed let the availaiblity check happen with infants
-        modifiedPaxConfigs = this.convertToAdult(modifiedPaxConfigs, passengerTypeMap.Infant);
-      }
-      return modifiedPaxConfigs;
-    };
-
-    this.parseDateRanges = dateRanges => {
-      const dateRangesResult = [];
-
-      if (!dateRanges) {
-        return dateRangesResult;
-      }
-
-      // Extract OptDateRange items from the dateRanges object
-      const optDateRanges = R.pathOr([], ['OptDateRange'], dateRanges);
-      const dateRangeArray = Array.isArray(optDateRanges) ? optDateRanges : [optDateRanges];
-
-      dateRangeArray.forEach(dateRange => {
-        const rateSets = R.pathOr({}, ['RateSets', 'RateSet'], dateRange);
-        const rateSet = Array.isArray(rateSets) ? rateSets[0] : rateSets;
-
-        const roomRates = R.pathOr({}, ['OptRate', 'RoomRates'], rateSet);
-        const extrasRates = R.pathOr({}, ['OptRate', 'ExtrasRates', 'ExtrasRate'], rateSet);
-
-        dateRangesResult.push({
-          startDate: dateRange.DateFrom,
-          endDate: dateRange.DateTo,
-          currency: dateRange.Currency,
-          priceCode: dateRange.PriceCode,
-          rateName: rateSet.RateName,
-          rateText: rateSet.RateText,
-          minSCU: rateSet.MinSCU,
-          maxSCU: rateSet.MaxSCU,
-          cancelHours: rateSet.CancelHours,
-          isClosed: rateSet.IsClosed,
-          scuCheckOverlapOnly: rateSet.ScuCheckOverlapOnly,
-          roomRates: {
-            singleRate: roomRates.SingleRate,
-            doubleRate: roomRates.DoubleRate,
-          },
-          extrasRates: {
-            adultRate: extrasRates.AdultRate,
-            childRate: extrasRates.ChildRate,
-            sequenceNumber: extrasRates.SequenceNumber,
-          },
-        });
-      });
-      return dateRangesResult;
-    };
-
-    /*
-      Get the stay results for an option based on start date, charge unit quantity, and room configs.
-
-      @param {string} optionId - The option ID
-      @param {string} hostConnectEndpoint - The HostConnect endpoint
-      @param {string} hostConnectAgentID - The agent ID
-      @param {string} hostConnectAgentPassword - The agent password
-      @param {Object} axios - The axios instance
-      @param {string} startDate - The start date
-      @param {number} chargeUnitQuantity - The number of charge units
-      @param {Object} roomConfigs - The room configurations
-      @param {string} displayRateInSupplierCurrency - Whether to see the availability rate in the supplier currency
-      @returns {Object} The stay results
-    */
-    this.getStayResults = async (
-      optionId,
-      hostConnectEndpoint,
-      hostConnectAgentID,
-      hostConnectAgentPassword,
-      axios,
-      startDate,
-      chargeUnitQuantity,
-      roomConfigs,
-      displayRateInSupplierCurrency,
-    ) => {
-      // The rate conversion flag: Y = convert to the currency of the agent, N = show in original currency
-      // From TP DOCS: If the value is Y then all rate information is converted to the currency associated
-      // with the specified agent. If it is false, no rate conversions are performed, and rates are returned
-      // in the currency in which they are stored. If RateConvert is not specified then whether currency
-      // conversion occurs or not is determined by a system default.
-      // Note: has no effect if R or S is not specified in Info.
-      const rateConvert = displayRateInSupplierCurrency && displayRateInSupplierCurrency.toUpperCase() === 'YES' ? 'N' : 'Y';
-      const getModel = checkType => ({
-        OptionInfoRequest: {
-          Opt: optionId,
-          Info: checkType,
-          DateFrom: startDate,
-          RateConvert: rateConvert, // for details see comments in header
-          SCUqty: (() => {
-            const num = parseInt(chargeUnitQuantity, 10);
-            if (isNaN(num) || num < 1) return 1;
-            return num;
-          })(),
-          RoomConfigs: roomConfigs,
-          AgentID: hostConnectAgentID,
-          Password: hostConnectAgentPassword,
-        },
-      });
-      // Always use G (General) & S (Stay & Availability) check types
-      const replyObj = await this.callTourplan({
-        model: getModel('GS'),
-        endpoint: hostConnectEndpoint,
-        axios,
-        xmlOptions: hostConnectXmlOptions,
-      });
-      const GSCheck = R.path(['OptionInfoReply', 'Option'], replyObj);
-      let OptStayResults = R.pathOr([], ['OptStayResults'], GSCheck);
-      if (!Array.isArray(OptStayResults)) OptStayResults = [OptStayResults];
-      return OptStayResults;
-
-      /* NOTES On the method data
-        If not rates, optionInfoReply is null, meaning it's not bookable
-        otherwise, example data:
-        {
-          optionInfoReply: {
-            Option: {
-              Opt: 'LHRTRDAVIDSHABTVC',
-              OptStayResults: {
-                AgentPrice: '51175',
-                Availability: 'RQ',
-                CancelHours: '96',
-                CommissionPercent: '0.00',
-                Currency: 'GBP',
-                PeriodValueAdds: {
-                  PeriodValueAdd: {
-                    DateFrom: '2024-08-10',
-                    DateTo: '2024-08-10',
-                    RateName: 'Std-Mon Sun',
-                    RateText: 'Std'
-                  }
-                },
-                RateId: '$NZD2016b81f44793096,4,DOUB,C5 A1INF KGB DOUB RONL',
-                RateName: 'Std-Mon Sun',
-                RateText: 'Std',
-                SaleFrom: '2023-12-20',
-                TotalPrice: '51175',
-                ExternalRateDetails: {
-                  ExtSupplierId: 'NZ000021',
-                  ExtOptionId: 'DOUB',
-                  ExtOptionDescr: 'Superior Room, 1 King Bed',
-                  ExtRatePlanCode: 'C5 A1INF KGB DOUB RONL',
-                  ExtRatePlanDescr: 'Room Only',
-                  ExtGuarantee: 'N',
-                },
-                RoomList: {
-                  RoomType: 'DB',
-                }
-              },
-              OptionNumber: '70461'
-            }
-          }
-        }
-      */
-      // const ACheckPass = (() => {
-      //   /*
-      //     FROM TP DOCS:
-      //     Each integer in the list gives the availability for one of the days in the range requested,
-      //     from the start date through to the end date. The integer values are to be interpreted as
-      //     follows:
-      //     Greater than 0 means that inventory is available, with the integer specifying the
-      //     number of units available. For options with a service type of Y , the inventory is in
-      //     units of rooms. For other service types, the inventory is in units of pax.
-      //     -1 Not available.
-      //     -2 Available on free sell.
-      //     -3 Available on request.
-      //     Note: A return value of 0 or something less than -3 is impossible.
-      //   */
-      //   const optAvail = parseInt(R.pathOr('-4', ['OptAvail'], ACheck), 10);
-      //   if (optAvail === -1) {
-      //     return {
-      //       available: false,
-      //     };
-      //   }
-      //   if (optAvail === -2) {
-      //     return {
-      //       available: true,
-      //       type: 'free sell',
-      //     };
-      //   }
-      //   if (optAvail === -3) {
-      //     return {
-      //       available: true,
-      //       type: 'on request',
-      //     };
-      //   }
-      //   if (optAvail > 0) {
-      //     return {
-      //       available: true,
-      //       type: 'inventory',
-      //       quantity: optAvail,
-      //     };
-      //   }
-      //   return {
-      //     available: false,
-      //   };
-      // })();
-    };
-
-    // eslint-disable-next-line arrow-body-style
-    this.getRatesObjectArray = (OptStayResults, markupPercentage = 0, OptStayResultsExtendedDates = []) => {
-      return OptStayResults.map(rate => {
-        const rateId = markupPercentage > 0 ? CUSTOM_RATE_ID_NAME : R.path(['RateId'], rate);
-        const currency = R.pathOr('', ['Currency'], rate);
-        // NOTE: Check if the value is in cents or not
-        const totalPrice = R.pathOr('', ['TotalPrice'], rate);
-        const agentPrice = R.pathOr('', ['AgentPrice'], rate);
-        let finalTotalPrice = Number(totalPrice);
-        let finalAgentPrice = Number(agentPrice);
-        let markupFactor = 1;
-        if (markupPercentage >= MIN_MARKUP_PERCENTAGE && markupPercentage <= MAX_MARKUP_PERCENTAGE) {
-          markupFactor = 1 + (Number(markupPercentage) / 100);
-        }
-
-        if (OptStayResultsExtendedDates.length > 0) {
-          const singleDayRate = OptStayResultsExtendedDates.find(rate2 => rate2.RateId === rate.RateId);
-          const totalPriceNoRatesDays = R.pathOr(0, ['TotalPrice'], singleDayRate);
-          const agentPriceNoRatesDays = R.pathOr(0, ['AgentPrice'], singleDayRate);
-          finalTotalPrice = Math.round(Number(totalPrice) + (Number(totalPriceNoRatesDays) * markupFactor));
-          finalAgentPrice = Math.round(Number(agentPrice) + (Number(agentPriceNoRatesDays) * markupFactor));
-        } else if (markupFactor > 1) {
-          finalTotalPrice = Math.round(Number(totalPrice) * markupFactor);
-          finalAgentPrice = Math.round(Number(agentPrice) * markupFactor);
-        }
-        const currencyPrecision = R.pathOr(2, ['currencyPrecision'], rate);
-        // Cancellations within this number of hours of service date incur a cancellation
-        // penalty of some sort.
-        const cancelHours = R.pathOr('', ['CancelHours'], rate);
-        /* Sample data: for cancel policies for the option id (not the external rate)
-          <CancelPolicies>
-            <CancelPenalty>
-                <Deadline>
-                    <OffsetUnitMultiplier>168</OffsetUnitMultiplier>
-                    <OffsetTimeUnit>Hour</OffsetTimeUnit>
-                    <DeadlineDateTime>2025-07-18T22:00:00Z</DeadlineDateTime>
-                </Deadline>
-                <InEffect>N</InEffect>
-                <LinePrice>1021200</LinePrice>
-                <AgentPrice>1021200</AgentPrice>
-            </CancelPenalty>
-            <CancelPenalty>
-                <Deadline>
-                    <OffsetUnitMultiplier>720</OffsetUnitMultiplier>
-                    <OffsetTimeUnit>Hour</OffsetTimeUnit>
-                    <DeadlineDateTime>2025-06-25T22:00:00Z</DeadlineDateTime>
-                </Deadline>
-                <InEffect>Y</InEffect>
-                <LinePrice>204240</LinePrice>
-                <AgentPrice>204240</AgentPrice>
-            </CancelPenalty>
-          </CancelPolicies>
-        */
-        let cancelPolicies = this.extractCancelPolicies(rate, ['CancelPolicies', 'CancelPenalty'], true)
-          .filter(policy => policy.inEffect === true);
-
-        let externalRateText = R.pathOr('', ['ExternalRateDetails', 'ExtOptionDescr'], rate);
-        const extRatePlanDescr = R.pathOr('', ['ExternalRateDetails', 'ExtRatePlanDescr'], rate);
-        if (extRatePlanDescr && !externalRateText.includes(extRatePlanDescr)) {
-          externalRateText = `${externalRateText} (${extRatePlanDescr})`;
-        }
-
-        /* Sample data: For external start times
-          <ExternalRateDetails>
-            <ExtStartTimes>
-              <ExtStartTime>2026-04-01T06:30:00</ExtStartTime>
-              <ExtStartTime>2026-04-01T07:30:00</ExtStartTime>
-              ...
-            </ExtStartTimes>
-          </ExternalRateDetails>
-        */
-        // Extract external start times
-        const extStartTimes = (() => {
-          // Note: Tourplan expects the start time in HH:MM format, so convert before sending to UI
-          const startTimes = R.pathOr([], ['ExternalRateDetails', 'ExtStartTimes', 'ExtStartTime'], rate);
-          if (!Array.isArray(startTimes)) {
-            // If single item, convert to array
-            return startTimes ? [{ startTime: startTimes.split('T')[1].substring(0, 5) }] : [];
-          }
-          return startTimes.map(startTime => ({ startTime: startTime.split('T')[1].substring(0, 5) }));
-        })();
-
-        /* Sample data: For external pickup and dropoff details
-          Address & ExtPointInfo are optional
-
-          <ExternalRateDetails>
-            <ExtPickupDetails>
-              <ExtPickupDetail>
-                  <ExtPointName>Adina/Vibe Waterfront</ExtPointName>
-                  <MinutesPrior>30</MinutesPrior>
-                  <Address>7 Kitchener Dr, Darwin</Address>
-                  <ExtPointInfo>Additional Info.</ExtPointInfo>
-              </ExtPickupDetail>
-              <ExtPickupDetail>
-                  <ExtPointName>Argus Hotel</ExtPointName>
-                  <MinutesPrior>30</MinutesPrior>
-                  <Address>13 Shepherd St, Darwin (Front of Hotel)</Address>
-                  <ExtPointInfo>Additional Info.</ExtPointInfo>
-              </ExtPickupDetail>
-              ...
-            </ExtDropoffDetails>
-          </ExternalRateDetails>
-        */
-        const extPickupDetails = (() => {
-          const pickupDetails = R.pathOr([], ['ExternalRateDetails', 'ExtPickupDetails', 'ExtPickupDetail'], rate);
-          if (!Array.isArray(pickupDetails)) {
-            // If single item, convert to array
-            return pickupDetails ? [pickupDetails] : [];
-          }
-          return pickupDetails.map(detail => ({
-            pointName: R.pathOr('', ['ExtPointName'], detail),
-            minutesPrior: R.pathOr('', ['MinutesPrior'], detail),
-            address: R.pathOr('', ['Address'], detail),
-            pointInfo: R.pathOr('', ['ExtPointInfo'], detail),
-          }));
-        })();
-
-        const extDropoffDetails = (() => {
-          const dropoffDetails = R.pathOr([], ['ExternalRateDetails', 'ExtDropoffDetails', 'ExtDropoffDetail'], rate);
-          if (!Array.isArray(dropoffDetails)) {
-            // If single item, convert to array
-            return dropoffDetails ? [dropoffDetails] : [];
-          }
-          return dropoffDetails.map(detail => ({
-            pointName: R.pathOr('', ['ExtPointName'], detail),
-            minutesPrior: R.pathOr('', ['MinutesPrior'], detail),
-            address: R.pathOr('', ['Address'], detail),
-            pointInfo: R.pathOr('', ['ExtPointInfo'], detail),
-          }));
-        })();
-
-        /* Sample data: for cancel policies for the external rate
-          <ExternalRateDetails>
-            <CancelPolicies>
-                <CancelPenalty>
-                    <PenaltyDescription>Cancellation 100% - within 24 hours or no notice</PenaltyDescription>
-                </CancelPenalty>
-                <CancelPenalty>
-                    <Deadline>
-                        <OffsetUnitMultiplier>2</OffsetUnitMultiplier>
-                        <OffsetTimeUnit>Day</OffsetTimeUnit>
-                    </Deadline>
-                    <PenaltyDescription>Day Tour Cancellation within 48hrs - 50%</PenaltyDescription>
-                </CancelPenalty>
-                <CancelPenalty>
-                    <PenaltyDescription>Day Tour Cancellation within 24hrs - 100%</PenaltyDescription>
-                </CancelPenalty>
-            </CancelPolicies>
-          </ExternalRateDetails>
-        */
-        if (cancelPolicies.length === 0) {
-          // If no cancel policies for the option, check the external rate
-          cancelPolicies = this.extractCancelPolicies(rate, ['ExternalRateDetails', 'CancelPolicies', 'CancelPenalty'], false);
-        }
-        /* Sample data: For additional details
-          <AdditionalDetails>
-            <AdditionalDetail>
-                <DetailName>Keywords</DetailName>
-                <DetailDescription>1|king|bed|classic|non|smoking</DetailDescription>
-            </AdditionalDetail>
-          </AdditionalDetails>
-        */
-        const additionalDetails = (() => {
-          const addDetails = R.pathOr([], ['ExternalRateDetails', 'AdditionalDetails', 'AdditionalDetail'], rate);
-          if (!Array.isArray(addDetails)) {
-            // If single item, convert to array
-            return addDetails ? [addDetails] : [];
-          }
-          return addDetails.map(detail => ({
-            detailName: R.pathOr('', ['DetailName'], detail),
-            detailDescription: R.pathOr('', ['DetailDescription'], detail),
-          }));
-        })();
-
-        return {
-          rateId,
-          currency,
-          totalPrice: finalTotalPrice,
-          agentPrice: finalAgentPrice,
-          currencyPrecision,
-          cancelHours,
-          externalRateText,
-          cancelPolicies,
-          startTimes: extStartTimes,
-          puInfoList: extPickupDetails.length ? extPickupDetails : [],
-          doInfoList: extDropoffDetails.length ? extDropoffDetails : [],
-          additionalDetails,
-        };
-      });
-    };
   }
 
   async validateToken({
@@ -1391,571 +501,17 @@ class BuyerPlugin {
 
   async searchProductsForItinerary({
     axios,
-    typeDefsAndQueries: {
-      itineraryProductTypeDefs,
-      itineraryProductQuery,
-    },
-    token: {
-      hostConnectEndpoint,
-      hostConnectAgentID,
-      hostConnectAgentPassword,
-      configuration,
-    },
-    payload: {
-      // single optionId or array of optionIds
-      optionId,
-      forceRefresh,
-      searchInput,
-      // lastUpdatedFrom is used to get options that were updated after a certain date in Tourplan
-      // example: lastUpdatedFrom: '2024-04-22 05:17:57.427Z'
-      lastUpdatedFrom,
-    },
+    token,
+    typeDefsAndQueries,
+    payload,
   }) {
-    /*
-      Pseudo
-      1. getServiceCodes -> [AC, BD]
-      2. for each serviceCode getoptions
-      3. convert them to ti2 products structure
-      4. merge all products from all serviceCodes
-    */
-    // getServices
-    const getServicesModel = {
-      GetServicesRequest: {
-        AgentID: hostConnectAgentID,
-        Password: hostConnectAgentPassword,
-      },
-    };
-    const getServicesReply = await this.callTourplan({
-      model: getServicesModel,
-      endpoint: hostConnectEndpoint,
+    return searchProductsForItinerary({
       axios,
-      xmlOptions: hostConnectXmlOptions,
+      token,
+      typeDefsAndQueries,
+      payload,
+      callTourplan: this.callTourplan.bind(this),
     });
-    let serviceCodes = R.pathOr([], ['GetServicesReply', 'TPLServices', 'TPLService'], getServicesReply);
-    if (!Array.isArray(serviceCodes)) serviceCodes = [serviceCodes];
-    serviceCodes = serviceCodes.map(s => s.Code);
-    let options = [];
-    await Promise.each(serviceCodes, async serviceCode => {
-      const getOptionsModel = {
-        OptionInfoRequest: {
-          Opt: `???${serviceCode}????????????`,
-          Info: 'G',
-          AgentID: hostConnectAgentID,
-          Password: hostConnectAgentPassword,
-          ...(lastUpdatedFrom ? {
-            LastUpdateFrom: lastUpdatedFrom,
-          } : {}),
-        },
-      };
-      const getOptionsReply = await this.callTourplan({
-        model: getOptionsModel,
-        endpoint: hostConnectEndpoint,
-        axios,
-        xmlOptions: hostConnectXmlOptions,
-      });
-      let thisOptions = R.pathOr([], ['OptionInfoReply', 'Option'], getOptionsReply);
-      // due to the new parser, single option will be returned as an object
-      // instead of an array
-      if (!Array.isArray(thisOptions)) thisOptions = [thisOptions];
-      console.log(`got ${thisOptions.length} options for serviceCode ${serviceCode}`);
-      options = options.concat(thisOptions);
-    });
-    if (!(options && options.length)) {
-      throw new Error('No products found');
-    }
-    const arrayOfOptionsGroupedBySupplierId = R.call(R.compose(
-      R.values,
-      R.groupBy(R.path(['OptGeneral', 'SupplierId'])),
-    ), options);
-    const products = await Promise.map(
-      arrayOfOptionsGroupedBySupplierId,
-      optionsGroupedBySupplierId => translateTPOption({
-        rootValue: {
-          optionsGroupedBySupplierId,
-        },
-        typeDefs: itineraryProductTypeDefs,
-        query: itineraryProductQuery,
-      }),
-      {
-        concurrency: 10,
-      },
-    );
-    return {
-      products,
-      productFields: [],
-      ...(searchInput || optionId ? {} : configuration),
-    };
-  }
-
-  /**
-   * Get and prepare availability configuration parameters (General and Date Ranges)
-   * @param {Object} params - Configuration parameters
-   * @returns {Object} Parsed configuration object
-   */
-  async getAvailabilityConfig({
-    optionId,
-    startDate,
-    chargeUnitQuantity,
-    paxConfigs,
-    hostConnectEndpoint,
-    hostConnectAgentID,
-    hostConnectAgentPassword,
-    axios,
-  }) {
-    const optionInfo = await this.getGeneralAndDateRangesInfo(optionId, hostConnectEndpoint, hostConnectAgentID, hostConnectAgentPassword, axios, startDate, chargeUnitQuantity);
-    const {
-      countChildrenInPaxBreak,
-      childrenAllowed,
-      countInfantsInPaxBreak,
-      infantsAllowed,
-      duration,
-      maxPaxPerCharge,
-      chargeUnit,
-      dateRanges,
-    } = optionInfo;
-
-    const modifiedPaxConfigs = this.getModifiedPaxConfigs(countChildrenInPaxBreak, childrenAllowed, countInfantsInPaxBreak, infantsAllowed, paxConfigs);
-    const roomConfigs = this.getRoomConfigs(modifiedPaxConfigs, true);
-
-    // Get the end date and message
-    const endDate = this.calculateEndDate(startDate, duration, chargeUnitQuantity);
-    const message = this.getOptionMessage(duration, chargeUnitQuantity, chargeUnit);
-
-    return {
-      roomConfigs,
-      endDate,
-      message,
-      dateRanges,
-      maxPaxPerCharge,
-    };
-  }
-
-  async getNoRatesAvailableError({
-    optionId,
-    hostConnectEndpoint,
-    hostConnectAgentID,
-    hostConnectAgentPassword,
-    axios,
-    endDate,
-    roomConfigs,
-  }) {
-    let errorMessage = GENERIC_AVALABILITY_CHK_ERROR_MESSAGE;
-    const immediateLastDateRange = await this.getImmediateLastDateRange(optionId, hostConnectEndpoint, hostConnectAgentID, hostConnectAgentPassword, axios, endDate, roomConfigs);
-    const dateTill = immediateLastDateRange ? immediateLastDateRange.endDate : null;
-    if (dateTill) {
-      const formattedDateTill = moment(dateTill).format(USER_FRIENDLY_DATE_FORMAT);
-      errorMessage = RATES_AVAILABLE_TILL_ERROR_TEMPLATE.replace('{dateTill}', formattedDateTill);
-    }
-    return {
-      bookable: false,
-      type: 'inventory',
-      rates: [],
-      message: errorMessage,
-    };
-  }
-
-  async searchAvailabilityForItinerary({
-    axios,
-    token: {
-      hostConnectEndpoint,
-      hostConnectAgentID,
-      hostConnectAgentPassword,
-      displayRateInSupplierCurrency,
-      customRatesEnableForQuotesAndBookings,
-      customRatesMarkupPercentage,
-      customRatesCalculateWithLastYearsRate,
-      customRatesExtendedBookingYears,
-    },
-    payload: {
-      optionId,
-      startDate,
-      /*
-      paxConfigs: [{ roomType: 'DB', adults: 2 }, { roomType: 'TW', children: 2 }]
-      */
-      paxConfigs,
-      // passengers,
-      /*
-        The number of second charge units required (second charge units are discussed
-        in the OptionInfo section). Should only be specified for options that have SCUs.
-        Defaults to 1.
-      */
-      chargeUnitQuantity,
-    },
-  }) {
-    // Get application configuration parameters for custom rates
-    const isBookingForCustomRatesEnabled = !!(customRatesEnableForQuotesAndBookings && customRatesEnableForQuotesAndBookings.toUpperCase() === 'YES');
-    const useLastYearRate = !!(customRatesCalculateWithLastYearsRate && customRatesCalculateWithLastYearsRate.toUpperCase() === 'YES');
-    // Assign default values when parameters are empty, null, undefined, or not a valid number between MIN_MARKUP_PERCENTAGE-MAX_MARKUP_PERCENTAGE
-    const markupPercentage = (() => {
-      const numValue = customRatesMarkupPercentage ? Number(customRatesMarkupPercentage) : DEFAULT_CUSTOM_RATE_MARKUP_PERCENTAGE;
-      return (numValue >= MIN_MARKUP_PERCENTAGE && numValue <= MAX_MARKUP_PERCENTAGE) ? numValue : DEFAULT_CUSTOM_RATE_MARKUP_PERCENTAGE;
-    })();
-    // Assign default values when parameters are empty, null, undefined, or not a valid number between MIN_EXTENDED_BOOKING_YEARS-MAX_EXTENDED_BOOKING_YEARS
-    const extendedBookingYears = (() => {
-      const years = customRatesExtendedBookingYears ? Number(customRatesExtendedBookingYears) : DEFAULT_CUSTOM_RATES_EXTENDED_BOOKING_YEARS;
-      return (years >= MIN_EXTENDED_BOOKING_YEARS && years <= MAX_EXTENDED_BOOKING_YEARS) ? years : DEFAULT_CUSTOM_RATES_EXTENDED_BOOKING_YEARS;
-    })();
-
-    // Get availability configuration parameters from Tourplan(General and Date Ranges)
-    const availabilityConfig = await this.getAvailabilityConfig({
-      optionId,
-      startDate,
-      chargeUnitQuantity,
-      paxConfigs,
-      hostConnectEndpoint,
-      hostConnectAgentID,
-      hostConnectAgentPassword,
-      axios,
-    });
-
-    const {
-      roomConfigs,
-      endDate,
-      message,
-      dateRanges,
-      maxPaxPerCharge,
-    } = availabilityConfig;
-
-    // Validate max pax per charge
-    const maxPaxPerChargeError = this.validateMaxPaxPerCharge({
-      roomConfigs,
-      maxPaxPerCharge,
-    });
-    if (maxPaxPerChargeError) {
-      return maxPaxPerChargeError;
-    }
-
-    const ratesRequiredTillDate = moment(startDate).add(chargeUnitQuantity - 1, 'days').format('YYYY-MM-DD');
-    const lastDateRangeEndDate = dateRanges.length > 0 ? moment(dateRanges[dateRanges.length - 1].endDate) : null;
-    const atLeastOneDateDoesNotHaveRatesAvailable = dateRanges.length === 0 ||
-      (lastDateRangeEndDate && lastDateRangeEndDate.isBefore(ratesRequiredTillDate));
-    if (atLeastOneDateDoesNotHaveRatesAvailable) {
-      // At least one date does not have rates available.
-      if (!isBookingForCustomRatesEnabled) {
-        // If custom rates are not enabled, return error
-        return this.getNoRatesAvailableError({
-          optionId,
-          hostConnectEndpoint,
-          hostConnectAgentID,
-          hostConnectAgentPassword,
-          axios,
-          endDate: endDate || startDate,
-          roomConfigs,
-        });
-      }
-
-      // Custom Rates are Enabled
-
-      // Step1 : Get rates for the days that have rates available
-      const noOfDaysRatesAvailable = lastDateRangeEndDate ? lastDateRangeEndDate.diff(moment(startDate), 'days') + 1 : 0;
-      let OptStayResults = [];
-      // Get stay rates for the given dates
-      if (noOfDaysRatesAvailable > 0) {
-        OptStayResults = await this.getStayResults(
-          optionId,
-          hostConnectEndpoint,
-          hostConnectAgentID,
-          hostConnectAgentPassword,
-          axios,
-          startDate,
-          noOfDaysRatesAvailable,
-          roomConfigs,
-          displayRateInSupplierCurrency,
-        );
-        const SCheckPass = Boolean(OptStayResults.length);
-        if (!SCheckPass) {
-          return {
-            bookable: Boolean(SCheckPass),
-            type: 'inventory',
-            rates: [],
-            message: GENERIC_AVALABILITY_CHK_ERROR_MESSAGE,
-          };
-        }
-      }
-
-      // Step2 : Get rates for the days that do not have any rates available
-      // This is done by getting the past rates based on the useLastYearRate flag.
-      // If true use last year's rates, otherwise use last available rates.
-      let pastDateAsStartDate = moment(startDate).subtract(1, 'year').format('YYYY-MM-DD');
-      if (!useLastYearRate) {
-        const immediateLastDateRange = await this.getImmediateLastDateRange(
-          optionId,
-          hostConnectEndpoint,
-          hostConnectAgentID,
-          hostConnectAgentPassword,
-          axios,
-          endDate || startDate,
-          roomConfigs,
-        );
-        if (!immediateLastDateRange) {
-          // If no immediate last date range found, return error
-          return {
-            bookable: false,
-            type: 'inventory',
-            rates: [],
-            message: GENERIC_AVALABILITY_CHK_ERROR_MESSAGE,
-          };
-        }
-
-        // If immediate last date range found, use the last rate's start date
-        pastDateAsStartDate = immediateLastDateRange.startDate;
-
-        // Check if the start date is beyond the permitted future booking years
-        const extendedBookingPermittedDate = moment(immediateLastDateRange.endDate).add(extendedBookingYears, 'years').format('YYYY-MM-DD');
-        if (moment(startDate).isAfter(extendedBookingPermittedDate)) {
-          return {
-            bookable: false,
-            type: 'inventory',
-            rates: [],
-            message: EXTENDED_BOOKING_YEARS_ERROR_TEMPLATE.replace('{lastRateEndDate}', immediateLastDateRange.endDate).replace('{extendedBookingYears}', extendedBookingYears),
-          };
-        }
-      }
-
-      // Format the success message for the custom rates
-      const customPeriodInfoMsg = useLastYearRate ? 'last year\'s rate.' : 'last available rate.';
-      const customRateInfoMsg = markupPercentage > 0 ? ` Custom rate applied, calculated using a markup on ${customPeriodInfoMsg}` : ` Custom rate applied with no markup on ${customPeriodInfoMsg}`;
-      const successMessage = message ? `${message}. ${customRateInfoMsg}` : `${customRateInfoMsg}`;
-
-      // Calculate the number of days to charge at the last rate
-      const daysToChargeAtLastRate = chargeUnitQuantity - noOfDaysRatesAvailable;
-
-      // Get stay rates
-      let OptStayResultsExtendedDates = await this.getStayResults(
-        optionId,
-        hostConnectEndpoint,
-        hostConnectAgentID,
-        hostConnectAgentPassword,
-        axios,
-        pastDateAsStartDate,
-        daysToChargeAtLastRate,
-        roomConfigs,
-        displayRateInSupplierCurrency,
-      );
-      const SCheckPass = Boolean(OptStayResultsExtendedDates.length);
-      if (SCheckPass) {
-        if (OptStayResults.length === 0) {
-          OptStayResults = OptStayResultsExtendedDates;
-          OptStayResultsExtendedDates = [];
-        }
-        return {
-          bookable: Boolean(SCheckPass),
-          type: 'inventory',
-          ...(endDate && SCheckPass ? { endDate } : {}),
-          ...(successMessage && SCheckPass ? { message: successMessage } : {}),
-          rates: this.getRatesObjectArray(OptStayResults, markupPercentage, OptStayResultsExtendedDates),
-        };
-      }
-      return {
-        bookable: false,
-        type: 'inventory',
-        rates: [],
-        message: GENERIC_AVALABILITY_CHK_ERROR_MESSAGE,
-      };
-    }
-
-    // Validate date ranges and room configurations
-    const dateRangesError = this.validateDateRanges({
-      dateRanges,
-      startDate,
-      chargeUnitQuantity,
-    });
-
-    if (dateRangesError) {
-      return dateRangesError;
-    }
-
-    // Get stay rates for the given dates
-    const OptStayResults = await this.getStayResults(
-      optionId,
-      hostConnectEndpoint,
-      hostConnectAgentID,
-      hostConnectAgentPassword,
-      axios,
-      startDate,
-      chargeUnitQuantity,
-      roomConfigs,
-      displayRateInSupplierCurrency,
-    );
-
-    const SCheckPass = Boolean(OptStayResults.length);
-    return {
-      bookable: Boolean(SCheckPass),
-      type: 'inventory',
-      ...(endDate && SCheckPass ? { endDate } : {}),
-      ...(message && SCheckPass ? { message } : {}),
-      rates: this.getRatesObjectArray(OptStayResults),
-    };
-  }
-
-  async addServiceToItinerary({
-    axios,
-    token: {
-      hostConnectEndpoint,
-      hostConnectAgentID,
-      hostConnectAgentPassword,
-    },
-    payload: {
-      quoteName,
-      itineraryOwner,
-      rateId,
-      quoteId,
-      // existingQuoteId,
-      // existingLineId,
-      optionId,
-      startDate,
-      reference,
-      /*
-      paxConfigs: [{ roomType: 'DB', adults: 2 }, { roomType: 'TW', children: 2 }]
-      */
-      paxConfigs,
-      /*
-        The number of second charge units required (second charge units are discussed
-        in the OptionInfo section). Should only be specified for options that have SCUs.
-        Defaults to 1.
-      */
-      chargeUnitQuantity,
-      extras,
-      startTime,
-      puInfo,
-      doInfo,
-      notes,
-      QB,
-      directHeaderPayload,
-      directLinePayload,
-      customFieldValues = [],
-      availCheckObj,
-    },
-  }) {
-    let pricing = null;
-    const tourplanServiceStatus = DEFAULT_TOURPLAN_SERVICE_STATUS;
-    const cfvPerService = customFieldValues.filter(f => f.isPerService && f.value)
-      .reduce((acc, f) => {
-        if (f.type === 'extended-option') {
-          acc[f.id] = f.value.value || f.value;
-        } else {
-          acc[f.id] = f.value;
-        }
-        return acc;
-      }, {});
-
-    if (availCheckObj && R.path(['rateId'], availCheckObj) === CUSTOM_RATE_ID_NAME) {
-      pricing = {
-        // ItemDescription: R.path(['Custom Rate'], availCheckObj),
-        CostCurrency: R.path(['currency'], availCheckObj),
-        CostConversionRate: R.pathOr(1, ['costConversionRate'], availCheckObj),
-        CostExclusive: R.path(['totalPrice'], availCheckObj),
-        // CostTax: R.path(['totalTax'], product),
-        AgentCurrency: R.path(['currency'], availCheckObj),
-        AgentExclusive: R.path(['agentPrice'], availCheckObj),
-        // AgentTax: R.path(['totalTax'], product),
-      };
-    }
-    // if external pickup and dropoff details are provided, use that info
-    // 1. If start time is provided send it in puTime
-    // 2. if extenral details are provided, send them in puRemark in the format:
-    //    (ExtPointName, ExtPointInfo, Address, Minutes prior)
-    // 3. the following shoud be sent:
-    //    puTime: '0930'
-    //    puRemark: 'Airport Pickup,Meet at arrivals hall,Airport Terminal 1,45,'
-    //    doTime: '1130' (Note: this is not used for external dropoff details)
-    //    doRemark: 'Hotel Dropoff,Drop at hotel entrance,456 Downtown Ave, City Center,15,',
-    let puTime = null;
-    let puRemark = null;
-    if (puInfo) {
-      if (puInfo.time || puInfo.location || puInfo.flightDetails) {
-        const puLocation = puInfo.location ? `${puInfo.location},` : '';
-        const puFlightDetails = puInfo.flightDetails ? `${puInfo.flightDetails},` : '';
-        if (puInfo.time && puInfo.time.replace(/\D/g, '')) {
-          puTime = puInfo.time.replace(/\D/g, '');
-        }
-        puRemark = this.escapeInvalidXmlChars(`${puLocation}${puFlightDetails}`);
-      } else if (puInfo.address || puInfo.pointName || puInfo.pointInfo || puInfo.minutesPrior) {
-        if (startTime) {
-          puTime = startTime.replace(/\D/g, '');
-        }
-        puRemark = this.escapeInvalidXmlChars(`${puInfo.pointName ? `${puInfo.pointName},` : ''}${puInfo.pointInfo ? `${puInfo.pointInfo},` : ''}${puInfo.address ? `${puInfo.address},` : ''}${puInfo.minutesPrior ? `${puInfo.minutesPrior},` : ''}`);
-      }
-    }
-
-    let doTime = null;
-    let doRemark = null;
-    if (doInfo) {
-      if (doInfo.time || doInfo.location || doInfo.flightDetails) {
-        const doLocation = doInfo.location ? `${doInfo.location},` : '';
-        const doFlightDetails = doInfo.flightDetails ? `${doInfo.flightDetails},` : '';
-        if (doInfo.time && doInfo.time.replace(/\D/g, '')) {
-          doTime = doInfo.time.replace(/\D/g, '');
-        }
-        doRemark = this.escapeInvalidXmlChars(`${doLocation}${doFlightDetails}`);
-      } else if (doInfo.address || doInfo.pointName || doInfo.pointInfo || doInfo.minutesPrior) {
-        // Note: There is no doTime for external dropoff details
-        doRemark = this.escapeInvalidXmlChars(`${doInfo.pointName ? `${doInfo.pointName},` : ''}${doInfo.pointInfo ? `${doInfo.pointInfo},` : ''}${doInfo.address ? `${doInfo.address},` : ''}${doInfo.minutesPrior ? `${doInfo.minutesPrior},` : ''}`);
-      }
-    }
-
-    const model = {
-      AddServiceRequest: {
-        AgentID: hostConnectAgentID,
-        Password: hostConnectAgentPassword,
-        // Note: Consult is optional but if provided is limited to 60 characters.
-        // In some systems is forced to uppercase. It is ignored if the Agent ID supplied is a sub-login.
-        // The field name in XML is Consult but on the TourPlan UI it is called Contact.
-        ...(itineraryOwner ? { Consult: this.escapeInvalidXmlChars(itineraryOwner).substring(0, 60) } : {}),
-        ...(quoteId ? {
-          ExistingBookingInfo: { BookingId: quoteId },
-        } : {
-          NewBookingInfo: {
-            Name: this.escapeInvalidXmlChars(quoteName),
-            QB: QB || 'Q',
-            ...(directHeaderPayload || {}),
-          },
-        }),
-        ...(puTime ? { puTime } : {}),
-        ...(puRemark ? { puRemark } : {}),
-        ...(doTime ? { doTime } : {}),
-        ...(doRemark ? { doRemark } : {}),
-        ...(extras && extras.filter(e => e.selectedExtra && e.selectedExtra.id).length ? {
-          ExtraQuantities: {
-            ExtraQuantityItem: extras.filter(e => e.selectedExtra && e.selectedExtra.id).map(e => ({
-              SequenceNumber: e.selectedExtra.id,
-              ExtraQuantity: e.quantity,
-            })),
-          },
-        } : {}),
-        Remarks: this.escapeInvalidXmlChars(notes).slice(0, 220),
-        Opt: optionId,
-        DateFrom: startDate,
-        RateId: rateId || 'Default',
-        ...(pricing ? { Pricing: pricing } : {}),
-        ...(pricing ? { TourplanServiceStatus: tourplanServiceStatus } : {}),
-        SCUqty: (() => {
-          const num = parseInt(chargeUnitQuantity, 10);
-          if (isNaN(num) || num < 1) return 1;
-          return num;
-        })(),
-        AgentRef: reference,
-        RoomConfigs: this.getRoomConfigs(paxConfigs),
-        ...(directLinePayload || {}),
-        ...(cfvPerService || {}),
-      },
-    };
-    const replyObj = await this.callTourplan({
-      model,
-      endpoint: hostConnectEndpoint,
-      axios,
-      xmlOptions: hostConnectXmlOptions,
-    });
-    return {
-      message: R.path(['AddServiceReply', 'Status'], replyObj)
-        === 'NO' ? SERVICE_CANNOT_BE_ADDED_ERROR_MESSAGE : '',
-      booking: {
-        id: R.path(['AddServiceReply', 'BookingId'], replyObj) || quoteId,
-        reference: R.path(['AddServiceReply', 'Ref'], replyObj),
-        linePrice: R.path(['AddServiceReply', 'Services', 'Service', 'LinePrice'], replyObj),
-        lineId: R.path(['AddServiceReply', 'ServiceLineId'], replyObj),
-      },
-    };
   }
 
   // eslint-disable-next-line class-methods-use-this
@@ -1998,102 +554,46 @@ class BuyerPlugin {
     };
   }
 
-  async searchItineraries({
-    token: {
-      hostConnectAgentID,
-      hostConnectAgentPassword,
-      hostConnectEndpoint,
-    },
+  async searchAvailabilityForItinerary({
     axios,
-    typeDefsAndQueries: {
-      itineraryBookingTypeDefs,
-      itineraryBookingQuery,
-    },
-    payload: {
-      purchaseDateStart,
-      purchaseDateEnd,
-      bookingId,
-      name,
-    },
+    token,
+    payload,
   }) {
-    const getPayload = (RequestType, RequestInput) => ({
-      model: {
-        [RequestType]: {
-          AgentID: hostConnectAgentID,
-          Password: hostConnectAgentPassword,
-          ...RequestInput,
-        },
-      },
-      endpoint: hostConnectEndpoint,
+    return searchAvailabilityForItinerary({
       axios,
-      xmlOptions: hostConnectXmlOptions,
+      token,
+      payload,
+      callTourplan: this.callTourplan.bind(this),
     });
-    const listBookingPayload = getPayload('ListBookingsRequest', {
-      EnteredDateFrom: purchaseDateStart || moment().subtract(6, 'month').format('YYYY-MM-DD'),
-      EnteredDateTo: purchaseDateEnd || moment().format('YYYY-MM-DD'),
-    });
-    let searchCriterias = [];
-    if (bookingId) {
-      searchCriterias = ['BookingId', 'Ref', 'AgentRef'].map(key => ({ [key]: this.escapeInvalidXmlChars(bookingId) }));
-    }
-    if (name) {
-      searchCriterias.push({ NameContains: this.escapeInvalidXmlChars(name) });
-    }
-    const allSearches = searchCriterias.length
-      ? searchCriterias.map(async keyObj => {
-        let reply;
-        try {
-          reply = await this.callTourplan(getPayload('ListBookingsRequest', keyObj));
-        } catch (err) {
-          if (err.includes && err.includes('Request failed with status code')) {
-            throw Error(err);
-          }
-          // if it's not server error, we just considered as no booking is found
-          // console.log('error in searchBooking', err);
-          reply = { ListBookingsReply: { BookingHeaders: { BookingHeader: [] } } };
-        }
+  }
 
-        // {"ListBookingsReply":
-        //  {"BookingHeaders":
-        //    {"BookingHeader":[
-        //        {"AgentRef":"2399181","BookingId":"314164","BookingStatus":"Quotation","BookingType":"F",
-        //          "Consult":null,"Currency":"GBP","EnteredDate":"2024-05-23","IsInternetBooking":"Y",
-        //          "Name":"Mr. Robert Slavonia x2 2399181","QB":"Q","Ref":"ALFI393309","TotalPrice":"583143",
-        //          "TravelDate":"2024-07-09"},{"AgentRef":"666666/1","BookingId":"315357",
-        //          "BookingStatus":"Quotation","BookingType":"F","Consult":null,"Currency":"GBP",
-        //          "EnteredDate":"2024-06-04","IsInternetBooking":"Y","Name":"Robert Guerrerio x3 2413898",
-        //          "QB":"Q","Ref":"ALFI393503","TotalPrice":"2309452","TravelDate":"2024-08-31"
-        //        }
-        //     ]}
-        // }}
-        return reply;
-      })
-      : [this.callTourplan(listBookingPayload)];
-    const replyObjs = await Promise.all(allSearches);
-    const bookingHeaders = R.flatten(replyObjs.map(o => R.pathOr([], ['ListBookingsReply', 'BookingHeaders', 'BookingHeader'], o)));
-    const bookings = await Promise.map(bookingHeaders, async bookingHeader => {
-      try {
-        const getBookingPayload = getPayload('GetBookingRequest', {
-          BookingId: R.prop('BookingId', bookingHeader),
-          ReturnAccountInfo: 'Y',
-          ReturnRoomConfigs: 'Y',
-        });
-        const bookingReply = await this.callTourplan(getBookingPayload);
-        const booking = R.path(['GetBookingReply'], bookingReply);
-        const newBooking = await translateItineraryBooking({
-          rootValue: booking,
-          typeDefs: itineraryBookingTypeDefs,
-          query: itineraryBookingQuery,
-        });
-        return newBooking;
-      } catch (err) {
-        console.error('error in searchBooking', err.message);
-        return null;
-      }
-    }, { concurrency: 10 });
-    return {
-      bookings: bookings.filter(b => b),
-    };
+  async addServiceToItinerary({
+    axios,
+    token,
+    payload,
+  }) {
+    return addServiceToItinerary({
+      axios,
+      token,
+      payload,
+      callTourplan: this.callTourplan.bind(this),
+    });
+  }
+
+  async searchItineraries({
+    axios,
+    typeDefsAndQueries,
+    token,
+    payload,
+  }) {
+    console.log('SACHIN searchItineraries CALLED');
+    return searchItineraries({
+      axios,
+      typeDefsAndQueries,
+      token,
+      payload,
+      callTourplan: this.callTourplan.bind(this),
+    });
   }
 }
